Added chess to navigator and connected home screen button to chessboard. Put reverse string into a util
