+++ mobile/src/components/Navigator.tsx
@@ -2,6 +2,7 @@ import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import React, { useContext } from "react";
 
 import { AuthContext } from "../contexts/AuthContext";
+import Chess from "../screens/Chess";
 import ForgotPassword from "../screens/ForgotPassword";
 import HomeScreen from "../screens/HomeScreen";
 import LoginScreen from "../screens/LoginScreen";
@@ -14,7 +15,10 @@ const Navigator = () => {
   return (
     <Stack.Navigator initialRouteName="Login" screenOptions={{ headerShown: false }}>
       {isLoggedIn ? (
-        <Stack.Screen name="HomeScreen" component={HomeScreen} />
+        <>
+          <Stack.Screen name="HomeScreen" component={HomeScreen} />
+          <Stack.Screen name="Chess" component={Chess} />
+        </>
       ) : (
         <>
           <Stack.Screen name="Login" component={LoginScreen} />
