+++ mobile/src/components/chess/Board.tsx
@@ -8,8 +8,8 @@ import { View, StyleSheet, Dimensions, Text } from "react-native";
 import Background from "./Background";
 import Gameover from "./Gameover";
 import Piece from "./Piece";
+import { reverseFenString } from "./util";
 
-const reverseString = (str) => str.split("").reverse().join("");
 const { width } = Dimensions.get("window");
 
 function useConst<T>(initialValue: T | (() => T)): T {
@@ -26,23 +26,6 @@ function useConst<T>(initialValue: T | (() => T)): T {
     };
   }
   return ref.current.value;
-}
-
-// Reverses the fen string to align with the perspective of the opposing player
-function reverseFenString(fen) {
-  const trailingText = fen.split("/").slice(-1)[0].split(" ").slice(1).join(" ");
-  const splitArray = fen
-    .split("/")
-    .slice(0, -1)
-    .concat([fen.split("/").slice(-1)[0].split(" ")[0]]);
-  for (let i = 0; i < splitArray.length; i++) {
-    splitArray[i] = reverseString(splitArray[i].split()[0]);
-  }
-  return splitArray
-    .reverse()
-    .slice(0, -1)
-    .concat(splitArray.slice(-1) + " " + trailingText)
-    .join("/");
 }
 
 const styles = StyleSheet.create({
