+++ mobile/src/screens/HomeScreen.tsx
@@ -1,3 +1,4 @@
+import { useNavigation } from "@react-navigation/native";
 import React, { useContext } from "react";
 import { Text, View } from "react-native";
 
@@ -6,12 +7,13 @@ import { UserContext } from "../contexts/UserContext";
 import { AppStylesheet } from "../styles/AppStylesheet";
 
 const HomeScreen = () => {
+  const navigation = useNavigation();
   const { userState } = useContext(UserContext);
 
   return (
     <View style={AppStylesheet.container}>
       <Text style={AppStylesheet.headerHomeScreen}>Welcome, {userState.firstName}</Text>
-      <Button text="Play Game" />
+      <Button text="Play Game" onPress={() => navigation.navigate("Chess")}/>
       <Text>14 Players Online</Text>
     </View>
   );
